refactor: cleaner separation
